fix(ci): unblock lint and migrations after dev merge
- Add 0066 merge migration for the two 0065 leaves introduced by the
  dev merge (notification_related_report + scoreauditlog_newcomer_boost)
- Move useMyReports into frontend/src/hooks so MyReports.tsx exports
  only components (react-refresh/only-export-components)
- Drop the redundant setError(null) at effect entry; initial state is
  already null and deps are [] (react-hooks/set-state-in-effect)
